# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type FormulaIngredientRole = typeof FormulaIngredientRole[keyof typeof FormulaIngredientRole];


export const FormulaIngredientRole = {
  top: 'top',
  heart: 'heart',
  base: 'base',
  modifier: 'modifier',
} as const;

export interface FormulaIngredient {
  materialId: number;
  materialName: string;
  percentage: number;
  grams: number;
  /**
     * Stock solution dilution percentage (100 = undiluted pure material)
     * @minimum 0
     * @maximum 100
     */
  dilution?: number;
  role: FormulaIngredientRole;
  allergenFlags?: string[];
}

export type FormulaIngredientInputRole = typeof FormulaIngredientInputRole[keyof typeof FormulaIngredientInputRole];


export const FormulaIngredientInputRole = {
  top: 'top',
  heart: 'heart',
  base: 'base',
  modifier: 'modifier',
} as const;

export interface FormulaIngredientInput {
  materialId: number;
  materialName: string;
  /** @minimum 0 */
  percentage: number;
  /** @minimum 0 */
  grams: number;
  /**
     * Stock solution dilution percentage (100 = undiluted pure material)
     * @minimum 0
     * @maximum 100
     */
  dilution?: number;
  role: FormulaIngredientInputRole;
  allergenFlags?: string[];
}

export type FormulaStatus = typeof FormulaStatus[keyof typeof FormulaStatus];


export const FormulaStatus = {
  draft: 'draft',
  resting: 'resting',
  approved: 'approved',
  archived: 'archived',
} as const;

export type FormulaSafetyStatus = typeof FormulaSafetyStatus[keyof typeof FormulaSafetyStatus];


export const FormulaSafetyStatus = {
  clear: 'clear',
  review: 'review',
  blocked: 'blocked',
} as const;

export type FormulaIfraStatus = typeof FormulaIfraStatus[keyof typeof FormulaIfraStatus];


export const FormulaIfraStatus = {
  within_limit: 'within_limit',
  review: 'review',
  exceeds_limit: 'exceeds_limit',
} as const;

export interface Formula {
  id: number;
  name: string;
  brief: string;
  status: FormulaStatus;
  concentration: number;
  totalMl: number;
  version: number;
  ingredients: FormulaIngredient[];
  safetyStatus: FormulaSafetyStatus;
  allergenCount: number;
  ifraStatus: FormulaIfraStatus;
  notes?: string;
  updatedAt: string;
  createdAt: string;
}

export type FormulaInputStatus = typeof FormulaInputStatus[keyof typeof FormulaInputStatus];


export const FormulaInputStatus = {
  draft: 'draft',
  resting: 'resting',
  approved: 'approved',
  archived: 'archived',
} as const;

export interface FormulaInput {
  /** @minLength 1 */
  name: string;
  brief: string;
  status: FormulaInputStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  concentration: number;
  /** @minimum 0 */
  totalMl: number;
  ingredients: FormulaIngredientInput[];
  notes?: string;
}

export type FormulaUpdateStatus = typeof FormulaUpdateStatus[keyof typeof FormulaUpdateStatus];


export const FormulaUpdateStatus = {
  draft: 'draft',
  resting: 'resting',
  approved: 'approved',
  archived: 'archived',
} as const;

export interface FormulaUpdate {
  /** @minLength 1 */
  name?: string;
  brief?: string;
  status?: FormulaUpdateStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  concentration?: number;
  /** @minimum 0 */
  totalMl?: number;
  ingredients?: FormulaIngredientInput[];
  notes?: string;
}

export type MaterialSafetyStatus = typeof MaterialSafetyStatus[keyof typeof MaterialSafetyStatus];


export const MaterialSafetyStatus = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Material {
  id: number;
  name: string;
  family: string;
  origin: string;
  /** @nullable */
  casNumber: string | null;
  allergens: string[];
  ifraCategory: string;
  ifraLimit: number;
  usageNotes: string;
  inStock: boolean;
  safetyStatus: MaterialSafetyStatus;
}

export interface DashboardSummary {
  formulaCount: number;
  materialCount: number;
  reviewCount: number;
  allergenCount: number;
  recentFormulas: Formula[];
  focusPrompt: string;
}

export interface CoachingMessageInput {
  /** @minLength 1 */
  message: string;
  /** @nullable */
  formulaId?: number | null;
  /** @nullable */
  formulaContext?: string | null;
}

export interface CoachingMessage {
  reply: string;
  suggestions: string[];
  cautions: string[];
}

export type ConversationMessageRole = typeof ConversationMessageRole[keyof typeof ConversationMessageRole];


export const ConversationMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ConversationMessage {
  id: number;
  conversationId: number;
  role: ConversationMessageRole;
  content: string;
  createdAt: string;
}

export interface Conversation {
  id: number;
  ownerId: string;
  title: string;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
}

export interface ConversationDetail {
  id: number;
  ownerId: string;
  title: string;
  createdAt: string;
  updatedAt: string;
  messages: ConversationMessage[];
}

export interface CreateConversationInput {
  /** @minLength 1 */
  title: string;
}

export interface SendConversationMessageInput {
  /** @minLength 1 */
  message: string;
  /** @nullable */
  formulaContext?: string | null;
}

export interface FormulaEvent {
  id: number;
  formulaId: number;
  formulaName: string;
  type: string;
  summary: string;
  createdAt: string;
}

export type ListFormulasParams = {
search?: string;
status?: ListFormulasStatus;
};

export type ListFormulasStatus = typeof ListFormulasStatus[keyof typeof ListFormulasStatus];


export const ListFormulasStatus = {
  draft: 'draft',
  resting: 'resting',
  approved: 'approved',
  archived: 'archived',
} as const;

export type ListMaterialsParams = {
search?: string;
};

export type GetActivityParams = {
limit?: number;
};