# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Perfumery workspace overview
  - name: formulas
    description: Formula library and revisions
  - name: materials
    description: Raw material library and safety data
  - name: coaching
    description: AI creative coaching
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get workspace summary
      responses:
        "200":
          description: Summary of the current perfumery workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /formulas:
    get:
      operationId: listFormulas
      tags: [formulas]
      summary: List formulas
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [draft, resting, approved, archived]
      responses:
        "200":
          description: Formula list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Formula"
    post:
      operationId: createFormula
      tags: [formulas]
      summary: Create a formula
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormulaInput"
      responses:
        "201":
          description: Created formula
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Formula"
  /formulas/{id}:
    get:
      operationId: getFormula
      tags: [formulas]
      summary: Get a formula
      parameters:
        - $ref: "#/components/parameters/FormulaId"
      responses:
        "200":
          description: Formula detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Formula"
        "404":
          description: Formula not found
    patch:
      operationId: updateFormula
      tags: [formulas]
      summary: Update a formula
      parameters:
        - $ref: "#/components/parameters/FormulaId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormulaUpdate"
      responses:
        "200":
          description: Updated formula
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Formula"
    delete:
      operationId: deleteFormula
      tags: [formulas]
      summary: Delete a formula
      parameters:
        - $ref: "#/components/parameters/FormulaId"
      responses:
        "204":
          description: Formula deleted
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List raw materials
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Material list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
  /coaching/messages:
    post:
      operationId: sendCoachingMessage
      tags: [coaching]
      summary: Ask the AI perfumery coach for guidance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoachingMessageInput"
      responses:
        "200":
          description: Coaching response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachingMessage"
  /conversations:
    get:
      operationId: listConversations
      tags: [coaching]
      summary: List coaching conversations
      responses:
        "200":
          description: Conversation list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [coaching]
      summary: Create a new coaching conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationDetail"
  /conversations/{conversationId}:
    get:
      operationId: getConversation
      tags: [coaching]
      summary: Get conversation with full message history
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      responses:
        "200":
          description: Conversation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationDetail"
        "404":
          description: Not found
    delete:
      operationId: deleteConversation
      tags: [coaching]
      summary: Delete a conversation
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      responses:
        "204":
          description: Deleted
  /conversations/{conversationId}/messages:
    post:
      operationId: sendConversationMessage
      tags: [coaching]
      summary: Send a message and get an AI response
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendConversationMessageInput"
      responses:
        "200":
          description: Updated conversation with new messages appended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationDetail"
  /formulas/{id}/events:
    get:
      operationId: getFormulaEvents
      tags: [formulas]
      summary: Get formula change log
      parameters:
        - $ref: "#/components/parameters/FormulaId"
      responses:
        "200":
          description: Formula events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FormulaEvent"
  /activity:
    get:
      operationId: getActivity
      tags: [dashboard]
      summary: Activity feed across all formulas
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: number
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FormulaEvent"
components:
  parameters:
    FormulaId:
      name: id
      in: path
      required: true
      schema:
        type: number
        minimum: 1
    ConversationId:
      name: conversationId
      in: path
      required: true
      schema:
        type: number
        minimum: 1
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    FormulaIngredient:
      type: object
      required: [materialId, materialName, percentage, grams, role]
      properties:
        materialId:
          type: number
        materialName:
          type: string
        percentage:
          type: number
        grams:
          type: number
        dilution:
          type: number
          minimum: 0
          maximum: 100
          default: 100
          description: Stock solution dilution percentage (100 = undiluted pure material)
        role:
          type: string
          enum: [top, heart, base, modifier]
        allergenFlags:
          type: array
          items:
            type: string
    FormulaIngredientInput:
      type: object
      required: [materialId, materialName, percentage, grams, role]
      properties:
        materialId:
          type: number
        materialName:
          type: string
        percentage:
          type: number
          minimum: 0
        grams:
          type: number
          minimum: 0
        dilution:
          type: number
          minimum: 0
          maximum: 100
          default: 100
          description: Stock solution dilution percentage (100 = undiluted pure material)
        role:
          type: string
          enum: [top, heart, base, modifier]
        allergenFlags:
          type: array
          items:
            type: string
    Formula:
      type: object
      required: [id, name, brief, status, concentration, totalMl, version, ingredients, updatedAt, createdAt, safetyStatus, allergenCount, ifraStatus]
      properties:
        id:
          type: number
        name:
          type: string
        brief:
          type: string
        status:
          type: string
          enum: [draft, resting, approved, archived]
        concentration:
          type: number
        totalMl:
          type: number
        version:
          type: number
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/FormulaIngredient"
        safetyStatus:
          type: string
          enum: [clear, review, blocked]
        allergenCount:
          type: number
        ifraStatus:
          type: string
          enum: [within_limit, review, exceeds_limit]
        notes:
          type: string
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    FormulaInput:
      type: object
      required: [name, brief, status, concentration, totalMl, ingredients]
      properties:
        name:
          type: string
          minLength: 1
        brief:
          type: string
        status:
          type: string
          enum: [draft, resting, approved, archived]
        concentration:
          type: number
          minimum: 0
          maximum: 100
        totalMl:
          type: number
          minimum: 0
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/FormulaIngredientInput"
        notes:
          type: string
    FormulaUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        brief:
          type: string
        status:
          type: string
          enum: [draft, resting, approved, archived]
        concentration:
          type: number
          minimum: 0
          maximum: 100
        totalMl:
          type: number
          minimum: 0
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/FormulaIngredientInput"
        notes:
          type: string
    Material:
      type: object
      required: [id, name, family, origin, casNumber, allergens, ifraCategory, ifraLimit, usageNotes, inStock, safetyStatus]
      properties:
        id:
          type: number
        name:
          type: string
        family:
          type: string
        origin:
          type: string
        casNumber:
          type: ["string", "null"]
        allergens:
          type: array
          items:
            type: string
        ifraCategory:
          type: string
        ifraLimit:
          type: number
        usageNotes:
          type: string
        inStock:
          type: boolean
        safetyStatus:
          type: string
          enum: [low, medium, high]
    DashboardSummary:
      type: object
      required: [formulaCount, materialCount, reviewCount, allergenCount, recentFormulas, focusPrompt]
      properties:
        formulaCount:
          type: number
        materialCount:
          type: number
        reviewCount:
          type: number
        allergenCount:
          type: number
        recentFormulas:
          type: array
          items:
            $ref: "#/components/schemas/Formula"
        focusPrompt:
          type: string
    CoachingMessageInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
        formulaId:
          type: ["number", "null"]
        formulaContext:
          type: ["string", "null"]
    CoachingMessage:
      type: object
      required: [reply, suggestions, cautions]
      properties:
        reply:
          type: string
        suggestions:
          type: array
          items:
            type: string
        cautions:
          type: array
          items:
            type: string
    ConversationMessage:
      type: object
      required: [id, conversationId, role, content, createdAt]
      properties:
        id:
          type: number
        conversationId:
          type: number
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    Conversation:
      type: object
      required: [id, ownerId, title, createdAt, updatedAt, messageCount]
      properties:
        id:
          type: number
        ownerId:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        messageCount:
          type: number
    ConversationDetail:
      type: object
      required: [id, ownerId, title, createdAt, updatedAt, messages]
      properties:
        id:
          type: number
        ownerId:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ConversationMessage"
    CreateConversationInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
    SendConversationMessageInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
        formulaContext:
          type: ["string", "null"]
    FormulaEvent:
      type: object
      required: [id, formulaId, formulaName, type, summary, createdAt]
      properties:
        id:
          type: number
        formulaId:
          type: number
        formulaName:
          type: string
        type:
          type: string
        summary:
          type: string
        createdAt:
          type: string
          format: date-time