# lib/api-zod/src/generated/types/conversationDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConversationMessage } from './conversationMessage';

export interface ConversationDetail {
  id: number;
  ownerId: string;
  title: string;
  createdAt: Date;
  updatedAt: Date;
  messages: ConversationMessage[];
}