# lib/api-zod/src/generated/types/formulaEvent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FormulaEvent {
  id: number;
  formulaId: number;
  formulaName: string;
  type: string;
  summary: string;
  createdAt: Date;
}