# lib/api-zod/src/generated/types/sendConversationMessageInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SendConversationMessageInput {
  /** @minLength 1 */
  message: string;
  /** @nullable */
  formulaContext?: string | null;
}